GH-977 Move ConnectionStatusMessage into DTOs package
