+++ aiida/src/test/java/energy/eddie/aiida/services/PermissionServiceTest.java
@@ -1,11 +1,11 @@
 package energy.eddie.aiida.services;
 
+import energy.eddie.aiida.dtos.ConnectionStatusMessage;
 import energy.eddie.aiida.dtos.PermissionDto;
 import energy.eddie.aiida.errors.*;
 import energy.eddie.aiida.models.permission.Permission;
 import energy.eddie.aiida.models.permission.PermissionStatus;
 import energy.eddie.aiida.repositories.PermissionRepository;
-import energy.eddie.aiida.streamers.ConnectionStatusMessage;
 import energy.eddie.aiida.streamers.StreamerManager;
 import nl.altindag.log.LogCaptor;
 import org.junit.jupiter.api.*;
